+++ src/components/mostrarPlan.vue
@@ -24,20 +24,38 @@
 <script>
 import { usePlanesStore } from '@/stores/planesStore';
 import api from '../axios.js'
+import Swal from 'sweetalert2';
 export default {
-    name:'mostrarPlan',
-    setup(){
+    name: 'mostrarPlan',
+    setup() {
         const planesStore = usePlanesStore();
 
         const eliminarPlan = async (id_plan) => {
-            try{
-                const response = await api.delete(`plan/delete/${id_plan}`);
-                console.log(response.data);
-            }catch(error){
+            try {
+                const result = await Swal.fire({
+                    title: 'Eliminar paciente',
+                    text: '¿Esta seguro de eliminar este paciente?',
+                    icon: 'question',
+                    showCancelButton: true,
+                    confirmButtonText: 'Eliminar',
+                    cancelButtonText: 'Cancelar',
+                    buttonsStyling: false,
+                    customClass: {
+                        confirmButton: 'btn btn-danger',
+                        cancelButton: 'btn btn-secondary'
+                    }
+                });
+                if (result.isConfirmed) {
+                    const response = await api.delete(`plan/delete/${id_plan}`);
+                    console.log(response.data);
+                    Swal.fire('¡Eliminado!', 'El archivo ha sido eliminado con éxito.', 'success');
+                }
+            } catch (error) {
                 console.log(error);
+                Swal.fire('Error', 'Hubo un problema al intentar eliminar el archivo.', 'error');
             }
         }
-        return {planesStore, eliminarPlan};
+        return { planesStore, eliminarPlan };
     }
 }
 </script>
